+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### New Features
 
+- Add "loop" static characteristic in Extractors
+
 ### Breaking Changes
 
 - Remove redundant `os_` parameter from `capa.loader.collect_metadata()`. External scripts calling this function must be updated to remove this argument. OS metadata now relies solely on the extractor yielding it via `extract_global_features()`. @mike-hunhoff #3076